add test, short_path
