+++ .gitignore
@@ -37,4 +37,10 @@ application/Example.ipynb
 application/.ipynb_checkpoints/*.*
 application/*.*
 
-
+src/.ipynb_checkpoints/bsp-checkpoint.dot
+src/.ipynb_checkpoints/Example-checkpoint.ipynb
+src/.ipynb_checkpoints/large-checkpoint.dot
+src/.ipynb_checkpoints/mein-checkpoint.dot
+src/.ipynb_checkpoints/output-checkpoint.png
+src/.ipynb_checkpoints/test-checkpoint.dot
+.gitignore
