--- src/.ipynb_checkpoints/bsp-checkpoint.dot
@@ -1,33 +1,0 @@
-digraph graphname {
-landscape=true;
- center=1;
- concentrate=true;
- 
- 1 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 2 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 3 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 4 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 5 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 6 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 7 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 8 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 9 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 10 [shape=circle,color=red,height=0.05,fontsize=40,margin=0,width=0.05];
- 2 -> 4 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 2 -> 7 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 2 -> 9 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 3 -> 5 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 4 -> 6 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 5 -> 1 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 6 -> 1 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 6 -> 3 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 7 -> 4 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 8 -> 1 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 8 -> 3 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 8 -> 5 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 8 -> 7 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 9 -> 5 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 9 -> 8 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 10 -> 2 [fontsize=40,arrowtype=normal,arrowsize=1.0];
- 10 -> 8 [fontsize=40,arrowtype=normal,arrowsize=1.0];
-}
