--- src/.ipynb_checkpoints/mein-checkpoint.dot
@@ -1,34 +1,0 @@
-graph G {
-  1 -> 6;
-  1 -> 7;
-  1 -> 9;
-  1 -> 10;
-  2 -> 3;
-  2 -> 5;
-  2 -> 8;
-  2 -> 9;
-  3 -> 2;
-  3 -> 4;
-  4 -> 3;
-  4 -> 8;
-  4 -> 9;
-  5 -> 2;
-  5 -> 9;
-  5 -> 10;
-  6 -> 1;
-  6 -> 8;
-  7 -> 1;
-  8 -> 2;
-  8 -> 4;
-  8 -> 6;
-  8 -> 9;
-  8 -> 10;
-  9 -> 1;
-  9 -> 2;
-  9 -> 4;
-  9 -> 5;
-  9 -> 8;
-  10 -> 1;
-  10 -> 5;
-  10 -> 8;
-}
