+++ test/runtests.jl
@@ -1,14 +1,33 @@
 using Network
+using Network # own!
+
 using Test
+using SparseArrays
+using Graphs
+using SimpleWeightedGraphs
+using DataStructures
+using LinearAlgebra
+using Plots
+using GraphPlot
+using DataStructures
+using ShowGraphviz
+using Cairo
+using Fontconfig
+using Random
+using Tokenize
+
+const testdir = dirname(@__FILE__)
+
+tests = [
+    "generators/test_generators",
+    "generators/test_graph_utils"
+]
 
 @testset "Network.jl" begin
-
-    # test creating network and 
-    m = 10
-    density = 0.2
-    net = Network.rand_network(m, density)
-    adj = Network.AdjacencyList(net.weights)
-    mat = Network.adjacency_matrix(adj)
-    @test mat == net.weights
-
+    for t in tests
+        tp = joinpath(testdir, "$(t).jl")
+        include(tp)
+    end
 end
+
+
